problema seed risolto

+++ database/migrations/2021_05_19_112241_create_travels_table.php
@@ -20,7 +20,7 @@ class CreateTravelsTable extends Migration
             $table->date('date');
             $table->tinyInteger('n_people');
             $table->smallInteger('price');
-            $table->tinyInteger('duration_days');
+            $table->smallInteger('duration_days');
             $table->timestamps();
         });
     }
